+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Add "Display summary" field to item enumeration data accordion. Refs UIIN-2740.
 * Display the 'Holdings detail' page after saving the holding changes. Fixes UIIN-2734.
 * *BREAKING* Bump up okapi interfaces for `pieces` (3.0). Refs UIIN-2761.
+* Action when Set record for deletion option is invoked. Refs UIIN-2595.
 
 ## [10.0.10](https://github.com/folio-org/ui-inventory/tree/v10.0.10) (2024-01-17)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v10.0.9...v10.0.10)